refactor: split storage.rs into module tree
Extract auxiliary types and free functions from the 7200-line storage.rs
into focused submodule files under src/storage/:

- events.rs: EventBus, PublishedAuditEvent, EventLogPage, EventLogPageOrder
- work_queue.rs: WorkQueuePromptProjection, WorkItemReadinessProjection,
  WorkItemCandidateClass, ActiveWaitConditionStates + ordering helpers
- recovery.rs: RecoverySnapshot, external_wait_recoverability_event
- activity.rs: FileActivityMarker, PollActivityMarker, file_activity_marker
- legacy_jsonl.rs: JSONL append/read/tail/migration helpers
- memory.rs: memory_index_agent_key

mod.rs retains AppStorage struct + impl + constants + tests.
All moved items are re-exported, preserving the public API.

Closes #1894
